adjustments and refactoring for exposure methods
